Updated layout and styles
Updated the layout and styles files after the suggestions of the mentor:
A style is a collection of attributes that specify the look and format
for a View or window. So if you have only one attribute there is no need
to define special style for it. You can directly apply this attribute in
the layout for view.